integrated the character type detail page into the characterId form, added a verbose prop to allow it to be lengthy when it needs to
