+++ src/components/character/types/TypeDetail.js
@@ -5,6 +5,7 @@ import ApiManager from "../../../modules/ApiManager";
 const TypeDetail = props => {
   const [subType, setSubType] = useState({});
   const id = props.subTypeId;
+  const verbose = props.verbose;
 
   const getSubType = () => {
     ApiManager.getSubTypeDetails(id)
@@ -13,7 +14,7 @@ const TypeDetail = props => {
 
   useEffect(()=>{
     getSubType();
-  }, [])
+  }, [id])
 
   return (
     <>
@@ -22,7 +23,7 @@ const TypeDetail = props => {
         <div className="header-container">
           {/* Having issues with this erroring out before full render */}
           {subType.characterType 
-            ? <h1> {subType.name}: {subType.characterType.name}</h1>
+            ? <h2> {subType.name}: {subType.characterType.name}</h2>
             : <></>   
           }
         </div>
@@ -35,65 +36,71 @@ const TypeDetail = props => {
             header="Purpose"
             description={<p>{subType.purpose}</p>}
           />
-          <Card
-            header="Aspects"
-            description={
-              <>
-                <p>
-                  {`Max Aspects: ${subType.maxAspects}`}
-                </p>
-                <p>
-                  {subType.aspectComment}
-                </p>
-              </>
-            }
-          />
-          <Card
-            header="Skills"
-            description={
-              <>
-                <p>
-                  {`Max Skill Rating: ${subType.maxSkillRating}`}
-                </p>
-                <p>
-                  {subType.skillRatingComment}
-                </p>
-                <p>
-                  {subType.skillChoiceComment}
-                </p>
-              </>
-            }
-          />
-          <Card
-            header="Stunts"
-            description={
-              <>
-                <p>
-                  {`Max Stunts: ${subType.maxStunts}`}
-                </p>
-                <p>
-                  {subType.stuntComment}
-                </p>
-              </>
-            }
-          />
-          <Card
-            header="Stress"
-            description={
-              <>
-                <p>
-                  {`Maximum Level of Stress Boxes: ${subType.stressMax}`}
-                </p>
-                <p>
-                  Stress Types: 
-                  {subType.bothStressTypes
-                    ? `Physical and Mental`
-                    : `One generic stress type`
+          {verbose === true 
+            ? <>
+                <Card
+                  header="Aspects"
+                  description={
+                    <>
+                      <p>
+                        {`Max Aspects: ${subType.maxAspects}`}
+                      </p>
+                      <p>
+                        {subType.aspectComment}
+                      </p>
+                    </>
                   }
-                </p>
+                />
+                <Card
+                  header="Skills"
+                  description={
+                    <>
+                      <p>
+                        {`Max Skill Rating: ${subType.maxSkillRating}`}
+                      </p>
+                      <p>
+                        {subType.skillRatingComment}
+                      </p>
+                      <p>
+                        {subType.skillChoiceComment}
+                      </p>
+                    </>
+                  }
+                />
+                <Card
+                  header="Stunts"
+                  description={
+                    <>
+                      <p>
+                        {`Max Stunts: ${subType.maxStunts}`}
+                      </p>
+                      <p>
+                        {subType.stuntComment}
+                      </p>
+                    </>
+                  }
+                />
+                <Card
+                  header="Stress"
+                  description={
+                    <>
+                      <p>
+                        {`Maximum Level of Stress Boxes: ${subType.stressMax}`}
+                      </p>
+                      <p>
+                        Stress Types: 
+                        {subType.bothStressTypes
+                          ? `Physical and Mental`
+                          : `One generic stress type`
+                        }
+                      </p>
+                    </>
+                  }
+                />
               </>
-            }
-          />
+            : <></>
+          }
+          
         </Card.Group>     
       </Container>
     </>
